requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
anthropic>=0.40.0
httpx>=0.27.0
schedule>=1.2.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
openpyxl>=3.1.0
